WIP Typescriptify IrcBridge
